Fixed >> Search box still display when navigating to detail screen from search results. #2
